+++ NEWS.md
@@ -1,3 +1,5 @@
+# NetFACS 0.5.0.9000
+
 # NetFACS 0.5.0
 
 ## Breaking changes